# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EpisodeIn": {
        "additionalProperties": false,
        "properties": {
          "highest_temperature_c": {
            "anyOf": [
              {
                "maximum": 45.0,
                "minimum": 25.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Temperature C"
          },
          "illness_description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illness Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeIn"
          },
          "trigger": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "trigger",
          "time"
        ],
        "title": "EpisodeIn",
        "type": "object"
      },
...
      "EpisodeOut": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "const": "fact",
            "default": "fact",
            "title": "Category",
            "type": "string"
          },
          "dose_count": {
            "default": 0,
            "title": "Dose Count",
            "type": "integer"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "highest_temperature_c": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Temperature C"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "illness_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illness Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "recovery_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Notes"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EpisodeStatus"
          },
          "symptom_count": {
            "default": 0,
            "title": "Symptom Count",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trigger",
          "status",
          "severity",
          "started_at",
          "ended_at",
          "timezone",
          "highest_temperature_c",
          "illness_description",
          "recovery_notes",
          "outcome",
          "notes"
        ],
        "title": "EpisodeOut",
        "type": "object"
      },
      "EpisodePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EpisodeOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "EpisodePage",
        "type": "object"
      },
...
      "EpisodeStatus": {
        "enum": [
          "open",
          "resolved",
          "escalated"
        ],
        "title": "EpisodeStatus",
        "type": "string"
      },
      "EpisodeUpdate": {
        "additionalProperties": false,
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventTimeIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "highest_temperature_c": {
            "anyOf": [
              {
                "maximum": 45.0,
                "minimum": 25.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Temperature C"
          },
          "illness_description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illness Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "recovery_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Notes"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpisodeStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "EpisodeUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InjectionIn": {
        "additionalProperties": false,
        "description": "Emergency injection. Only the clinical essentials are required (SAFE-23).",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 10000.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,11}0*$)\\d{0,6}\\.\\d{0,4}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "contact_notified": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Notified"
          },
          "emergency_services_called": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Services Called"
          },
          "episode_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "injected_by": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injected By"
          },
          "injection_site": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injection Site"
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "route": {
            "$ref": "#/components/schemas/Route",
            "default": "intramuscular"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeIn"
          },
          "transported_to_hospital": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transported To Hospital"
          },
          "unit": {
            "$ref": "#/components/schemas/DoseUnit",
            "default": "mg"
          }
        },
        "required": [
          "medication_id",
          "amount",
          "time"
        ],
        "title": "InjectionIn",
        "type": "object"
      },
      "InjectionOut": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "category": {
            "const": "fact",
            "default": "fact",
            "title": "Category",
            "type": "string"
          },
          "emergency_services_called": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Services Called"
          },
          "episode_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "injected_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injected By"
          },
          "injection_site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injection Site"
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceOut"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeOut"
          },
          "transported_to_hospital": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transported To Hospital"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "medication_id",
          "amount",
          "unit",
          "route",
          "time",
          "provenance",
          "injection_site",
          "reason",
          "injected_by",
          "response",
          "emergency_services_called",
          "transported_to_hospital",
          "episode_id"
        ],
        "title": "InjectionOut",
        "type": "object"
      },
      "InjectionPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InjectionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/InjectionOut"
            },
            "title": "Revisions",
            "type": "array"
          }
        },
        "required": [
          "items",
          "revisions",
          "page"
        ],
        "title": "InjectionPage",
        "type": "object"
      },
...
      "PageMetadata": {
        "additionalProperties": false,
        "description": "Shared bounded pagination state for growing owner-scoped collections.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 1.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "PageMetadata",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HealthCurve",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/emergency-injections": {
      "get": {
        "operationId": "list_injections_api_v1_emergency_injections_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjectionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Injections",
        "tags": [
          "episodes"
        ]
      },
      "post": {
        "description": "Log an emergency injection.\n\nDeliberately depends on nothing but the database: no AI call, no integration, no\nbackground job (SAFE-23). Everything beyond medication, amount, unit, and time is\noptional, because a partial record now beats a complete record later.",
        "operationId": "log_injection_api_v1_emergency_injections_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjectionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Injection",
        "tags": [
          "episodes"
        ]
      }
    },
...
    "/api/v1/stress-episodes": {
      "get": {
        "operationId": "list_episodes_api_v1_stress_episodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EpisodeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Open Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episodes",
        "tags": [
          "episodes"
        ]
      },
      "post": {
        "operationId": "create_episode_api_v1_stress_episodes_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Episode",
        "tags": [
          "episodes"
        ]
      }
    },
    "/api/v1/stress-episodes/{episode_id}": {
      "patch": {
        "description": "Update an open episode.\n\nAn episode is a container the owner curates while it is happening, so it is\ngenuinely mutable -- unlike the recorded facts inside it, which are corrected by\nsupersession.",
        "operationId": "update_episode_api_v1_stress_episodes__episode_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Episode",
        "tags": [
          "episodes"
        ]
      }
    },
...
  }
}